app/requirements.txt
PyYAML
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Session
pyodbc
cryptography
requests
requests